Make io_service::schedule() noexcept.
In the case that a schedule_operation could not be queued to the
I/O completion port using PostQueuedCompletionStatus() it now
pushes the operation onto a lock-free linked-list of overflow
operations.
The event loop will check for these overflow operations and try
to requeue them to the I/O completion port at a later time.

This allows us to declare io_service::schedule_impl() noexcept
and thus the corresponding schedule_operation::await_suspend()
operation noexcept.