Fix incorrect code documentation for task.
There was a section left over from before lazy_task was renamed
to task and the eager task was removed that was contrasting
lazy_task with the eager task. This comparison is no longer
necessary.

Removed the warnings about stack-overflow as these issues have
been worked around (albeit at cost of an atomic operation).