Simplify shared_task reference counting.
No longer decrement the reference count in final_suspend()
as it's not needed for a lazy shared_task where there is
always guaranteed to be one awaiting shared_task.
This also fixes a potential memory leak of the coroutine
frame if a shared_task is never awaited.

Modified condition used for decrementing ref-count to
work around potential codegen bug in Clang (PR35068).